feat!: expand type system to support double, string, and bool
- Integrated double as the primary numeric type for all calculations.
- Updated tokenizer and parser to handle decimal literals and boolean
values.

+++ Morphyn.Core/Program.cs
@@ -66,7 +66,7 @@ namespace Morphyn.Core
                 {
                     DateTime currentTime = DateTime.Now;
                     // Calculate fps
-                    int dtMs = (int)(currentTime - lastTime).TotalMilliseconds;
+                    double dtMs = (currentTime - lastTime).TotalMilliseconds;
                     lastTime = currentTime;
 
                     foreach (var entity in context.Entities.Values)
